rename utils to delta_utils
